+++ src/index.ts
@@ -3,3 +3,4 @@ export { uniquifyArray, uniquifyObjectsArray } from './functions/uniquify-array'
 export { objectsArrayToObject } from './functions/objects-array-to-object';
 export { sortArrayByObjectProperty } from './functions/sort-array-by-object-property';
 export { groupArrayByObjectProperty } from './functions/group-array-by-object-property';
+export { getNumberEnumKeys, getNumberEnumValues } from './functions/get-number-enum-values';